REVERT: adds support for -a and specifying file
The old revert has been split into PerforceRevert and
PerforceRevertInternal. The old one functions almost the same with the
addition of vargs to specify additional revert arguments namely -a. It
will call PerforceRevertInternal to do the heavy lifting but will also
handle the refreshing of the file to reset the Readonly flag

PerforceRevertInternal now does the system call and command building.
However it takes the additional argument filename to allow the target
file to be set. This allows it to be called from autocommands where %
might not report the filename.